remove embedded boehm gc sources, use distro copy instead
boehm GC is a highly problematic package, since what it tries to achieve
is not possible in standard C, so it must use all kinds of nasty
platform-specific tricks to get start offset of stack and so on.

for this reason, some or most distributions apply custom settings
or even patches to make it work.

these tweaks are not available when embedding the source of boehm into
the tree and hoping that it will turn out to work well...

windows users are advised to install the version provided by mingw
package manager, while mac os users can use homebrew's version.

closes #6
closes #9
closes #4