fix ram allocation scheme of LEG parser
the original code used to have a tiny patch on top of peg 0.1.4
overriding all calls to malloc/realloc via a YYMALLOC/YYREALLOC
macro, so that all parts use the same GC (probably boehm).

re-applying this change was overlooked when we updated to latest
peg. meanwhile peg supplies its own YY_MALLOC which we now use
instead.

this caused odd issues such as "TypeError" message when running
the strings.rb test on linux distributions based on musl, which
is generally more sensible to undefined behaviour.