grammar.leg: fix bug preventing zero-termination of some strings
the STRING_PUSH macro does not zero-terminate strings, so we must
use the pushed strings with TrString_new which takes a length argument
and takes care of terminating the produced string, unlike TrString_new2
which uses strlen() on the input string, requiring it to be zero-
terminated.
this happened to work in the past by accident, because boehm gc hands
out only zeroed memory.